+++ app/packages/app/src/Network.tsx
@@ -37,8 +37,6 @@ export const NetworkRenderer = ({ makeRoutes }) => {
   const { context, environment } = fos.useRouter(makeRoutes, []);
 
   const isModalActive = Boolean(useRecoilValue(fos.modal));
-  // is this necessary for color modal too?
-  const isCustomizeColorModalActive = useRecoilValue(fos.activeColorField);
 
   useEffect(() => {
     document.body.classList.toggle("noscroll", isModalActive);
@@ -46,16 +44,6 @@ export const NetworkRenderer = ({ makeRoutes }) => {
       .getElementById("modal")
       ?.classList.toggle("modalon", isModalActive);
   }, [isModalActive]);
-  // is this necessary for color modal too?
-  useEffect(() => {
-    document.body.classList.toggle(
-      "noscroll",
-      Boolean(isCustomizeColorModalActive)
-    );
-    document
-      .getElementById("colorModal")
-      ?.classList.toggle("modalon", Boolean(isCustomizeColorModalActive));
-  }, [isCustomizeColorModalActive]);
 
   return <Network environment={environment} context={context} />;
 };
